refactor(vet): 병원 견적서 조회 컴포넌트 변경
