+++ apps/vet/src/interfaces/estimate/index.ts
@@ -1,0 +1,17 @@
+export interface VetEstimateGeneralList {
+  id: number;
+  imageUrl: string;
+  nickname: string;
+  proposal: 'GENERAL';
+  significant: string;
+  reservedDate: string;
+}
+
+export interface VetEstimateDesignationList {
+  id: number;
+  imageUrl: string;
+  nickname: string;
+  proposal: 'DESIGNATION';
+  significant: string;
+  reservedDate: string;
+}
