+++ apps/vet/src/models/estimate/index.ts
@@ -1,27 +1,11 @@
+import { VetEstimateDesignationList, VetEstimateGeneralList } from '~/interfaces/estimate';
+
 export interface GetVetEstimateGeneralListResponse {
   estimates: VetEstimateGeneralList[];
 }
 
-export interface VetEstimateGeneralList {
-  id: number;
-  imageUrl: string;
-  nickname: string;
-  proposal: 'GENERAL';
-  significant: string;
-  reservedDate: string;
-}
-
 export interface GetVetEstimateDesignationListResponse {
   estimates: VetEstimateDesignationList[];
-}
-
-export interface VetEstimateDesignationList {
-  id: number;
-  imageUrl: string;
-  nickname: string;
-  proposal: 'DESIGNATION';
-  significant: string;
-  reservedDate: string;
 }
 
 export interface GetVetEstimateDetailRequestParams {
