+++ packages/services/package.json
@@ -42,6 +42,7 @@
     "@emotion/react": "^11.13.5",
     "@tanstack/react-query": "^5.61.0",
     "axios": "^1.7.7",
+    "dayjs": "^1.11.13",
     "dotenv": "^16.4.7",
     "nanoid": "^5.0.9",
     "react": "^18.3.1",
