+++ packages/services/src/components/estimate-list/card/index.tsx
@@ -13,6 +13,8 @@ import {
   specialsNot,
   detailContainer,
 } from './index.styles';
+import dayjs from 'dayjs';
+import 'dayjs/locale/ko';
 
 interface EstimateGeneralListType {
   id: number;
@@ -22,21 +24,6 @@ interface EstimateGeneralListType {
   significant: string;
   reservedDate: string;
   onDetailClick: () => void;
-}
-
-function formatDate(dateString: string): string {
-  const date = new Date(dateString);
-
-  const year = date.getFullYear();
-  const month = String(date.getMonth() + 1).padStart(2, '0');
-  const day = String(date.getDate()).padStart(2, '0');
-  const hours = String(date.getHours()).padStart(2, '0');
-  const minutes = String(date.getMinutes()).padStart(2, '0');
-
-  const weekdays = ['일', '월', '화', '수', '목', '금', '토'];
-  const weekday = weekdays[date.getDay()];
-
-  return `${year}.${month}.${day}(${weekday}) ${hours}:${minutes}`;
 }
 
 export function Card({
@@ -52,10 +39,6 @@ export function Card({
     event.currentTarget.onerror = null;
     event.currentTarget.src = '';
   };
-
-  const formattedDate = reservedDate
-    ? formatDate(new Date(reservedDate.toString()).toISOString())
-    : formatDate(reservedDate);
 
   return (
     <div css={wrapper} onClick={onDetailClick}>
@@ -82,7 +65,7 @@ export function Card({
             {significant ? significant : '특이사항 없음'}
           </p>
           <Text typo="body11" color="gray500">
-            {formattedDate}
+            {dayjs(reservedDate).locale('ko').format('YYYY.MM.DD(ddd) • HH:mm')}
           </Text>
         </div>
       </div>
